refactor(orders): clean up redundant files and refactor to use redux hooks
