+++ application/src/components/index.js
@@ -1,10 +1,8 @@
 import Login from './login/login';
 import Main from './main/main';
 import Nav from './nav/nav';
-import OrderForm from './order-form/orderForm';
-import OrderFormHook from './order-form-hook/order-form';
+import OrderForm from './order-form-hook/orderForm';
 import Template from './common/template';
 import ViewOrders from './view-orders/viewOrders';
-import ViewOrdersHook from './view-orders-hook/viewOrders';
 
-export { Login, OrderForm, OrderFormHook, Main, Nav, Template, ViewOrders, ViewOrdersHook };
+export { Login, OrderForm, Main, Nav, Template, ViewOrders };
