+++ application/src/components/order-form-hook/orderForm.js
@@ -1,28 +1,26 @@
 import React, { useState } from 'react';
+import { Template } from '..';
 import { useSelector } from 'react-redux';
-import { Template } from '../../components';
 import { SERVER_IP } from '../../private';
 import './orderForm.css';
 
-const ADD_ORDER_URL = `${SERVER_IP}/api/add-order`;
+const ADD_ORDER_URL = `${SERVER_IP}/api/add-order`
 
-export default function OrderForm(props) {
+export default OrderForm = () => {
     const [orderItem, setOrderItem] = useState("");
     const [quantity, setQuantity] = useState("1");
 
-    const menuItemChosen = (event) => setOrderItem(event.value);
-    const menuQuantityChosen = (event) => setQuantity(event.value);
+    const userEmail = useSelector(state => state.email);
 
-    const auth = useSelector((state) => state.auth);
-
-    const submitOrder = () => {
+    const submitOrder = (event) => {
+        event.preventDefault();
         if (orderItem === "") return;
         fetch(ADD_ORDER_URL, {
             method: 'POST',
             body: JSON.stringify({
                 order_item: orderItem,
-                quantity,
-                ordered_by: auth.email || 'Unknown!',
+                quantity: quantity,
+                ordered_by: userEmail || 'Unknown!',
             }),
             headers: {
                 'Content-Type': 'application/json'
@@ -40,7 +38,7 @@ export default function OrderForm(props) {
                     <label className="form-label">I'd like to order...</label><br />
                     <select 
                         value={orderItem} 
-                        onChange={(event) => menuItemChosen(event)}
+                        onChange={(event) => setOrderItem(event.target.value)}
                         className="menu-select"
                     >
                         <option value="" defaultValue disabled hidden>Lunch menu</option>
@@ -50,7 +48,7 @@ export default function OrderForm(props) {
                         <option value="Chili Con Carne">Chili Con Carne</option>
                     </select><br />
                     <label className="qty-label">Qty:</label>
-                    <select value={quantity} onChange={(event) => menuQuantityChosen(event)}>
+                    <select value={quantity} onChange={(event) => setQuantity(event.target.value)}>
                         <option value="1">1</option>
                         <option value="2">2</option>
                         <option value="3">3</option>
@@ -58,9 +56,9 @@ export default function OrderForm(props) {
                         <option value="5">5</option>
                         <option value="6">6</option>
                     </select>
-                    <button type="button" className="order-btn" onClick={() => submitOrder()}>Order It!</button>
+                    <button type="button" className="order-btn" onClick={submitOrder}>Order It!</button>
                 </form>
             </div>
         </Template>
-    )
+    );
 }
