--- application/src/components/view-orders/viewOrders.css
@@ -1,51 +1,0 @@
-.view-order-container {
-    font-family: "Optima", serif;
-    min-height: 100px;
-    border-bottom: solid 1px #999;
-}
-
-.view-order-container p {
-    margin: 0;
-}
-
-.view-order-container button {
-    margin-left: 20px;
-}
-
-.view-order-left-col {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.view-order-middle-col {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.view-order-right-col {
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-}
-
-@media only screen and (max-width: 768px) {
-    .view-order-left-col {
-        justify-content: center;
-    }
-
-    .view-order-middle-col {
-        margin-bottom: 20px;
-    }
-
-    .view-order-right-col {
-        padding-bottom: 20px;
-        justify-content: space-around;
-    }
-
-    .view-order-container button {
-        margin: 0;
-    }
-}
