+++ application/src/router/appRouter.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import { Main, Login, OrderFormHook, ViewOrdersHook } from '../components';
+import { Main, Login, OrderForm, ViewOrdersHook } from '../components';
 
 const AppRouter = (props) => {
   return (
     <Router>
       <Route path="/" exact component={Main} />
       <Route path="/login" exact component={Login} />
-      <Route path="/order" exact component={OrderFormHook} />
+      <Route path="/order" exact component={OrderForm} />
       <Route path="/view-orders" exact component={ViewOrdersHook} />
     </Router>
   );
